+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 -   Reformatted the code to conform to the Prettier code style.
 -   Fixed mounts of typos.
+-   Added a new type `IntRange` in `utils.d.ts` and applied it to those parameters that accept a range of integers.
+-   Restricted the value range of `levelManager.cameraMode` to `0 | 1 | 2`.
+-   Restricted the value range of `levelManager.cameraDirection` to `0 | 1 | 2 | 3`.
+-   Restricted the value range of the parameter `which` of `uiElement.onPointerClick()` to `0 | 1 | 2 `.
 
 ## v1.0.6 (2025-05-29)